+++ lib/controllers/useSignInByAuth0.ts
@@ -2,7 +2,7 @@ import { signIn as nextAuthSignIn } from "next-auth/react";
 import { useRouter } from "next/router";
 import { useCallback } from "react";
 import { OmitFirstParameters } from "../../types/util";
-import { signInByAuth0 } from "../usecase/_signInByAuth0";
+import { signInByAuth0 } from "../usecases/_signInByAuth0";
 
 /**
  * @description
