+++ lib/controllers/useSignOut.ts
@@ -2,7 +2,7 @@ import { signOut as nextAuthSignOut } from "next-auth/react";
 import { useRouter } from "next/router";
 import { useCallback } from "react";
 import { OmitFirstParameters } from "../../types/util";
-import { signOut } from "../usecase/_signOut";
+import { signOut } from "../usecases/_signOut";
 
 /**
  * @description
