+++ lib/usecases/interface.ts
